Model Configurations

MITgcm_config

The data structure that enables MITgcm.jl is called MITgcm_config.

MITgcm.MITgcm_config — Type

MITgcm_config()

Concrete type of AbstractModelConfig for MITgcm (as part of the ClimateModels.jl interface for MITgcm) which contains

    model :: String = "MITgcm"
    configuration :: String = ""
    inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    channel :: Channel{Any} = Channel{Any}(10) 
    folder :: String = tempdir()
    ID :: UUID = UUIDs.uuid4()

and can be constructed using keywords as follows

unknown_config=MITgcm_config(configuration="unknown")

ECCO solutions

Global ocean model configurations used in NASA's ECCO ocean state estimation program.

ECCO4 : Forget et al., 2015 (doi:10.5194/gmd-8-3071-2015)

OCCA2 : Forget, 2024 (under review).

MITgcm.setup_ECCO4! — Function

setup_ECCO4!(config::MITgcm_config)

Setup method for ECCO4 and OCCA2 solutions.

fil=joinpath("examples","configurations","OCCA2.toml")
MC=MITgcm_config(inputs=read_toml(fil))
setup(MC)
build(MC)
MITgcm_launch(MC)

MITgcm/Verification

The MITgcm/verification folder is found in the MITgcm source code. It provides a suite of basic configurations used by model developers.

To use them, you want to download the MITgcm source code. For example :

using MITgcm
MITgcm_download()
MITgcm_path[1]

"/home/runner/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/MITgcm_test"

To get the list of standard model configurations in MITgcm/verification/ :

verification_experiments()

8-element Vector{MITgcm_config}:
   ID            = d79b4b92-e2d2-4a3d-8d59-4918b5971604
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/d79b4b92-e2d2-4a3d-8d59-4918b5971604
  log subfolder = /tmp/d79b4b92-e2d2-4a3d-8d59-4918b5971604/log

   ID            = 83cc9b6f-c9c2-4c00-9cfd-d8f0823f71cd
  model         = MITgcm
  configuration = advect_cs
  run folder    = /tmp/83cc9b6f-c9c2-4c00-9cfd-d8f0823f71cd
  log subfolder = /tmp/83cc9b6f-c9c2-4c00-9cfd-d8f0823f71cd/log

   ID            = 1c274d52-e9ed-45f3-b42d-349291297529
  model         = MITgcm
  configuration = advect_xy
  run folder    = /tmp/1c274d52-e9ed-45f3-b42d-349291297529
  log subfolder = /tmp/1c274d52-e9ed-45f3-b42d-349291297529/log

   ID            = 7f27f77a-036c-42c2-9724-ff5679a2829c
  model         = MITgcm
  configuration = aim.5l_cs
  run folder    = /tmp/7f27f77a-036c-42c2-9724-ff5679a2829c
  log subfolder = /tmp/7f27f77a-036c-42c2-9724-ff5679a2829c/log

   ID            = dc88e2dc-e84a-4d8f-88f1-cc337bb335d4
  model         = MITgcm
  configuration = flt_example
  run folder    = /tmp/dc88e2dc-e84a-4d8f-88f1-cc337bb335d4
  log subfolder = /tmp/dc88e2dc-e84a-4d8f-88f1-cc337bb335d4/log

   ID            = 80167869-30fb-4c23-bb81-ab9722c96810
  model         = MITgcm
  configuration = hs94.cs-32x32x5
  run folder    = /tmp/80167869-30fb-4c23-bb81-ab9722c96810
  log subfolder = /tmp/80167869-30fb-4c23-bb81-ab9722c96810/log

   ID            = a6e2959f-2280-485b-91dd-6e128eb97048
  model         = MITgcm
  configuration = tutorial_global_oce_biogeo
  run folder    = /tmp/a6e2959f-2280-485b-91dd-6e128eb97048
  log subfolder = /tmp/a6e2959f-2280-485b-91dd-6e128eb97048/log

   ID            = 7123e506-55aa-4c40-aa32-462182c5317f
  model         = MITgcm
  configuration = tutorial_held_suarez_cs
  run folder    = /tmp/7123e506-55aa-4c40-aa32-462182c5317f
  log subfolder = /tmp/7123e506-55aa-4c40-aa32-462182c5317f/log

To choose one :

MITgcm_config(configuration="MLAdjust")

  ID            = 84bd71ec-1258-4a16-b7db-427a33c7c28e
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/84bd71ec-1258-4a16-b7db-427a33c7c28e
  log subfolder = /tmp/84bd71ec-1258-4a16-b7db-427a33c7c28e/log

Functionalities

MITgcm.setup_verification! — Function

setup_verification!(config::MITgcm_config)

Setup method for verification experiments.

MITgcm.verification_experiments — Function

verification_experiments()

Get list of all most-standard configurations of MITgcm and return as an Array of MITgcm_config

exps=verification_experiments()

verification_experiments(nam::String)

Get one configurations of MITgcm and return as a MITgcm_config

advect_xy=verification_experiments("advect_xy")

MITgcm.MITgcm_download — Function

MITgcm_download()

Download default, compact version of MITgcm from zenodo.

MITgcm.MITgcm_path — Constant

MITgcm_path

Path to a MITgcm folder. MITgcm_path[1] should generally be used. MITgcm_path[2] is mostly meant to facilitate comparisons between e.g. MITgcm releases when needed.