Manual

MITgcmTools.jl provides a suite of tools for running MITgcm, analyzing its output, and/or modifying its inputs.

Reading MITgcm Files

MITgcmTools.read_mdsio — Function

read_mdsio(datafile)

Read a MITgcm mdsio file (".data" binary + ".meta" text pair), and return as an Array

p="./hs94.cs-32x32x5/run/"
x=read_mdsio(p*"surfDiag.0000000020.002.001.data")
y=read_mdsio(p*"pickup.ckptA.002.001.data")
z=read_mdsio(p*"T.0000000000.002.001.data")

read_mdsio(pth::String,fil::String)

Read a MITgcm's MDSIO files (".data" binary + ".meta" text pair), combine, and return as an Array

p="./hs94.cs-32x32x5/run/"
x=read_mdsio(p,"surfDiag.0000000020")
y=read_mdsio(p,"pickup.ckptA")
z=read_mdsio(p,"T.0000000000")

MITgcmTools.read_meta — Function

read_meta(metafile)

Read a MITgcm metadata file, parse it, and return as a NamedTuple

p="./hs94.cs-32x32x5/run/"
meta=read_meta(p*"surfDiag.0000000020.002.001.meta")
pairs(meta)
meta.dimList

read_meta(pth::String,fil::String)

Read a MITgcm metadata files, parse them, and return as an array of NamedTuple

p="./hs94.cs-32x32x5/run/"
meta=read_meta(p,"surfDiag.0000000020")
pairs(meta[end])
[meta[i].dimList for i in 1:length(meta)]

MITgcmTools.read_mnc — Function

read_mnc(pth::String,fil::String,var::String)

MITgcmTools.GridLoad_mdsio — Function

GridLoad_mdsio(myexp::MITgcm_config)

MITgcmTools.GridLoad_mnc — Function

GridLoad_mnc(γ::gcmgrid)

GridLoad_mnc(myexp::MITgcm_config)

MITgcmTools.read_available_diagnostics — Function

read_available_diagnostics(fldname::String; filename="available_diagnostics.log")

Get the information for a particular variable fldname from the available_diagnostics.log text file generated by MITgcm.

MITgcmTools.read_bin — Function

read_bin(fil::String,kt::Union{Int,Missing},kk::Union{Int,Missing},prec::DataType,mygrid::gcmgrid)

Read model output from binary file and convert to MeshArray. Other methods:

read_bin(fil::String,prec::DataType,mygrid::gcmgrid)
read_bin(fil::String,mygrid::gcmgrid)

MITgcmTools.read_flt — Function

read_flt(dirIn::String,prec::DataType)

Read displacements from MITgcm/pkg/flt output file into a DataFrame.

MITgcmTools.read_nctiles — Function

read_nctiles(fileName,fldName,mygrid)

Read model output from NCTiles file and convert to MeshArray instance.

mygrid=GridSpec("LatLonCap")
fileName="nctiles_grid/GRID"
Depth=read_nctiles(fileName,"Depth",mygrid)
hFacC=read_nctiles(fileName,"hFacC",mygrid)
hFacC=read_nctiles(fileName,"hFacC",mygrid,I=(:,:,1))

MITgcmTools.MITgcm_namelist — Type

MITgcm_namelist(groups,params)

Data structure representing a MITgcm namelist file, such as data.pkg, which contains

    groups :: Array{Symbol,1} = Array{Symbol,1}(undef, 0)
    params :: Array{OrderedDict{Symbol,Any},1} = Array{OrderedDict{Symbol,Any},1}(undef, 0)

with model parameters (params) being organized in groups as done in the files.

using MITgcmTools
fil=joinpath(MITgcm_path[1],"verification","advect_xy","run","data")
nml=read_namelist(fil)
MITgcm_namelist(nml.groups,nml.params)
MITgcm_namelist(groups=nml.groups,params=nml.params)
MITgcm_namelist(groups=nml.groups)

MITgcmTools.read_namelist — Function

read_namelist(fil)

Read a MITgcm namelist file, parse it, and return as a NamedTuple

using MITgcmTools
testreport("advect_xy")
fil=joinpath(MITgcm_path[1],"verification","advect_xy","run","data")
namelist=read_namelist(fil)

MITgcmTools.write_namelist — Function

write_namelist(fil)

Save a MITgcm namelist file. In the example below, one is read from file, modified, and then saved to a new file using write_namelist.

using MITgcmTools
fil=joinpath(MITgcm_path[1],"verification","advect_xy","run","data")
nml=read_namelist(fil)
write_namelist(fil*"_new",namelist)

or

nml=read(fil,MITgcm_namelist())
write(fil*"_new",nml)

Model Configurations

The verification_experiments() function provides a list of the most-standard MITgcm configurations that can all be run either in batch mode or interactively using MITgcmTools.jl.

The MITgcm_path variable points to where MITgcm is compiled. Interactive / reactive notebooks are found in the examples/ folder (e.g. MITgcm_run.jl seen just below).

Missing docstring.

Missing docstring for MITgcm_path. Check Documenter's build log for details.

MITgcmTools.verification_experiments — Function

verification_experiments()

Get list of all most-standard configurations of MITgcm and return as an Array of MITgcm_config

using MITgcmTools
exps=verification_experiments()

MITgcmTools.scan_rundir — Function

scan_rundir(pth::String)

Scan a MITgcm run directory and standard output text file ("output.txt" or "STDOUT.0000") and return a NamedTuple of collected information (various formats)

Initially, the output looked like (grid=gr,packages=pac,params_time=par1,params_grid=par2,completed=co)

Running MITgcm

The recommended, simple, method to run the model is via the climate model interface (see the ClimateModels.jl interface section and docs@ClimateModels.jl for detail). The MITgcm_launch function can be used to run a MITgcm configuration after setting up the MITgcm_config. Using this interface facilitates operations like compiling and setting up a temporary folder to run the model. Key functions, incl. the climate model interface, are documented further down in the docs.

The verification_experiments function provides a list of the most-standard MITgcm configurations that can all be run either in batch mode or interactively. The MITgcm_path variable points to where MITgcm is compiled. Interactive / reactive notebooks are found in the examples/ folder (e.g. run_MITgcm.jl seen just below).

ClimateModels.jl interface

MITgcmTools.MITgcm_config — Type

MITgcm_config()

Concrete type of AbstractModelConfig for MITgcm which contains

    model :: String = "MITgcm"
    configuration :: String = ""
    options :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    status :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    channel :: Channel{Any} = Channel{Any}(10) 
    folder :: String = tempdir()
    ID :: UUID = UUIDs.uuid4()

and with defaults that can be constructed as follows for example

using MITgcmTools
tmp=MITgcm_config()

exps=verification_experiments()
exps[end]

(part of the climate model interface as specialized for MITgcm)

ClimateModels.build — Function

build(config::MITgcm_config)

Build the model using genmake2, make depend, and make. The first two link all code files, headers, etc in the build/ folder before compiling the model

(part of the climate model interface as specialized for MITgcm)

build(config::MITgcm_config,options::String)

Build the model using genmake2, make depend, and make unless otherwise specified via options. The genmake2 and make depend commands link all code files, headers, etc in the build/ folder before make compiles the model.

(part of the climate model interface as specialized for MITgcm)

ClimateModels.compile — Function

compile(config::MITgcm_config)

Compile the model using make in build/ that has already been setup.

(part of the climate model interface as specialized for MITgcm)

ClimateModels.setup — Function

setup(config::MITgcm_config)

Create a run/ folder and link everything there as needed to be ready to run model as normally done for most-standard MITgcm configurations (incl. prepare_run and mitgcmuv). Call git_log_init(config) to setup git tracker and put!(config.channel,MITgcm_launch) to be executed via launch(config) later.

(part of the climate model interface as specialized for MITgcm)

MITgcmTools.MITgcm_launch — Function

MITgcm_launch(config::MITgcm_config)

Go to run/ folder and effectively call mitgcmuv > output.txt

(part of the climate model interface as specialized for MITgcm)

ClimateModels.clean — Function

clean(config::MITgcm_config)

Cancel any remaining task (config.channel) and clean up the run directory (via rm).

(part of the climate model interface as specialized for MITgcm)

Format conversions

MITgcmTools.findtiles — Function

findtiles(ni::Int,nj::Int,mygrid::gcmgrid)
findtiles(ni::Int,nj::Int,grid::String="LatLonCap",GridParentDir="../inputs/GRID_LLC90/")

Return a MeshArray map of tile indices, mytiles["tileNo"], for tile size ni,nj and extract grid variables accordingly.

MITgcmTools.cube2compact — Function

cube2compact(x::Array)

Reshape from e.g. size (192, 32, 5) in cube format to (32, 192, 5) in compact format.

MITgcmTools.compact2cube — Function

compact2cube(x::Array)

Reshape from e.g. size (32, 192, 5) in cube format to (192, 32, 5) in compact format.

MITgcmTools.convert2array — Function

convert2array(fld::MeshArray)

Convert MeshArray to Array (or vice versa otherwise)

MITgcmTools.convert2gcmfaces — Function

convert2gcmfaces(fld::MeshArray)

Convert mitgcm output to MeshArray (or vice versa otherwise)

Formulae etc

MITgcmTools.SeaWaterDensity — Function

SeaWaterDensity(Θ,Σ,Π,Π0)

Compute potential density (ρP), in situ density (ρI), and density referenced to PREF (Π0 in decibars) from potential temperature (Θ in °C), salinity (Σ in psu) and pressure (Π in decibars) according to the UNESCO / Jackett & McDougall 1994 equation of state.

Credits: code based on a Matlab implementation by B. Ferron Reference: https://www.jodc.go.jp/info/iocdoc/UNESCOtech/059832eb.pdf Check value: ρI = 1041.83267kg/m^3 for Θ=3°Celcius, Σ=35psu, Π=3000dbar

(ρP,ρI,ρR) = SeaWaterDensity(3.,35.5,3000.)
isapprox(ρI,1041.83267, rtol=1e-6)

MITgcmTools.MixedLayerDepth — Function

MixedLayerDepth(Θ,Σ,Δ,mthd)

Compute mixed layer depth from potential temperature (Θ in °C), salinity (Σ in psu) and depth (Δ in method) according to various formulas (mthd == "BM", "Suga", "Kara"). Inputs must be dense vectors without any missing value (or NaN, etc).

D=collect(0.0:1.0:500.0); tmp=(1.0.-tanh.(5*(-1 .+ 2/D[end]*D)));
T=2.0 .+ 8.0*tmp; S=34.0 .+ 0.5*tmp;
(ρP,ρI,ρR) = SeaWaterDensity(T,S,D);

mld=MixedLayerDepth(T,S,D,"BM"); isapprox(mld,134.0)

using Plots
plot(ρP,-D,w=2,label="Potential Density",ylabel="Depth")
plot!(vec([ρP[1] ρP[end]]),-fill(mld,2),label="Mixed Layer Depth",w=2,c="black",s=:dash)

API Reference

MITgcmTools.MITgcm_config

MITgcmTools.MITgcm_namelist

ClimateModels.build

ClimateModels.clean

ClimateModels.compile

ClimateModels.setup

MITgcmTools.GridLoad_mdsio

MITgcmTools.GridLoad_mnc

MITgcmTools.MITgcm_launch

MITgcmTools.MixedLayerDepth

MITgcmTools.SeaWaterDensity

MITgcmTools.compact2cube

MITgcmTools.convert2array

MITgcmTools.convert2gcmfaces

MITgcmTools.cube2compact

MITgcmTools.findtiles

MITgcmTools.read_available_diagnostics

MITgcmTools.read_bin

MITgcmTools.read_flt

MITgcmTools.read_mdsio

MITgcmTools.read_meta

MITgcmTools.read_mnc

MITgcmTools.read_namelist

MITgcmTools.read_nctiles

MITgcmTools.scan_rundir

MITgcmTools.verification_experiments

MITgcmTools.write_namelist