Model Configurations

MITgcm_config

The data structure that enables MITgcm.jl is called MITgcm_config.

MITgcm.MITgcm_config — Type

MITgcm_config()

Concrete type of AbstractModelConfig for MITgcm (as part of the ClimateModels.jl interface for MITgcm) which contains

    model :: String = "MITgcm"
    configuration :: String = ""
    inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    channel :: Channel{Any} = Channel{Any}(10) 
    folder :: String = tempdir()
    ID :: UUID = UUIDs.uuid4()

and can be constructed using keywords as follows

unknown_config=MITgcm_config(configuration="unknown")

ECCO solutions

Global ocean model configurations used in NASA's ECCO ocean state estimation program.

ECCO4 : Forget et al., 2015 (doi:10.5194/gmd-8-3071-2015)

OCCA2 : Forget, 2024 (under review).

MITgcm.setup_ECCO4! — Function

setup_ECCO4!(config::MITgcm_config)

Setup method for ECCO4 and OCCA2 solutions.

using MITgcm
fil=joinpath("examples","configurations","OCCA2.toml")
MC=MITgcm_config(inputs=read_toml(fil))
setup(MC)
build(MC)
launch(MC)

MITgcm.ECCO4_inputs.get_list — Function

get_list(; write_file=false)

Create a list of Dataverse folders for ECCOv4r2. If write_file=true then write to file joinpath(tempdir(),"Dataverse_list.csv")/

get_list(list1::DataFrame,name::String)

Create a list of Dataverse files from folder with specified name.

list1=ECCO4_inputs..get_list()
nam1="surface forcing fields"
list2=ECCO4_inputs..get_list(list1,nam1)

MITgcm.ECCO4_inputs.get_files — Function

get_files(list1::DataFrame,nam1::String)

Create a list of Dataverse files from folder with specified name.

list1=ECCO4_inputs..get_list()
nam1="model initialization"
ECCO4_inputs..get_files(list1,nam1,tempname())

MITgcm.ECCO4_testreport.compute — Function

compute(pth0)

@everywhere begin
 using MITgcm
 using ECCO4_testreport.SharedArrays
end

report=ECCO4_testreport.compute("run")

MITgcm.ECCO4_testreport.compare — Function

compare(A::DataFrame,B::DataFrame)

using ClimateModels.DataFrames, ClimateModels.CSV
ref_file="test/testreport_baseline2.csv"
ref=CSV.read(ref_file,DataFrame)

using MITgcm
ECCO4_testreport.compare(report,ref)

compare(A::DataFrame,B::DataFrame,v::AbstractString)

MITgcm/Verification

The MITgcm/verification folder is found in the MITgcm source code. It provides a suite of basic configurations used by model developers.

To use them, you want to download the MITgcm source code. For example :

using MITgcm
MITgcm_download()
MITgcm_path[1]

"/home/runner/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/MITgcm_test"

To get the list of standard model configurations in MITgcm/verification/ :

verification_experiments()

8-element Vector{MITgcm_config}:
   ID            = dfbf1f50-a1ce-4864-b280-0b88051bf3b9
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/dfbf1f50-a1ce-4864-b280-0b88051bf3b9
  log subfolder = /tmp/dfbf1f50-a1ce-4864-b280-0b88051bf3b9/log

   ID            = f7867d09-6195-410c-b68a-b8586d2fd4cf
  model         = MITgcm
  configuration = advect_cs
  run folder    = /tmp/f7867d09-6195-410c-b68a-b8586d2fd4cf
  log subfolder = /tmp/f7867d09-6195-410c-b68a-b8586d2fd4cf/log

   ID            = 195002ea-a123-4f85-a7cb-5512e3bc73b0
  model         = MITgcm
  configuration = advect_xy
  run folder    = /tmp/195002ea-a123-4f85-a7cb-5512e3bc73b0
  log subfolder = /tmp/195002ea-a123-4f85-a7cb-5512e3bc73b0/log

   ID            = 781ace33-21ce-4694-9063-e363fb838af5
  model         = MITgcm
  configuration = aim.5l_cs
  run folder    = /tmp/781ace33-21ce-4694-9063-e363fb838af5
  log subfolder = /tmp/781ace33-21ce-4694-9063-e363fb838af5/log

   ID            = 7540a225-0762-4150-a045-aea77ba7d047
  model         = MITgcm
  configuration = flt_example
  run folder    = /tmp/7540a225-0762-4150-a045-aea77ba7d047
  log subfolder = /tmp/7540a225-0762-4150-a045-aea77ba7d047/log

   ID            = 4da82c65-1124-4f7d-8b07-b05fa000f4c0
  model         = MITgcm
  configuration = hs94.cs-32x32x5
  run folder    = /tmp/4da82c65-1124-4f7d-8b07-b05fa000f4c0
  log subfolder = /tmp/4da82c65-1124-4f7d-8b07-b05fa000f4c0/log

   ID            = 4dd7dc43-b567-40ad-985e-d355197363e3
  model         = MITgcm
  configuration = tutorial_global_oce_biogeo
  run folder    = /tmp/4dd7dc43-b567-40ad-985e-d355197363e3
  log subfolder = /tmp/4dd7dc43-b567-40ad-985e-d355197363e3/log

   ID            = 1aa36182-f639-4eb5-aa1a-a625598ac5f7
  model         = MITgcm
  configuration = tutorial_held_suarez_cs
  run folder    = /tmp/1aa36182-f639-4eb5-aa1a-a625598ac5f7
  log subfolder = /tmp/1aa36182-f639-4eb5-aa1a-a625598ac5f7/log

To choose one :

MITgcm_config(configuration="MLAdjust")

  ID            = 992d1a0d-d0f5-4676-92b6-398acbe059ed
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/992d1a0d-d0f5-4676-92b6-398acbe059ed
  log subfolder = /tmp/992d1a0d-d0f5-4676-92b6-398acbe059ed/log

Functionalities

MITgcm.setup_verification! — Function

setup_verification!(config::MITgcm_config)

Setup method for verification experiments.

MITgcm.verification_experiments — Function

verification_experiments()

Get list of all most-standard configurations of MITgcm and return as an Array of MITgcm_config

exps=verification_experiments()

verification_experiments(nam::String)

Get one configurations of MITgcm and return as a MITgcm_config

advect_xy=verification_experiments("advect_xy")

MITgcm.MITgcm_download — Function

MITgcm_download()

Download default, compact version of MITgcm from zenodo.

MITgcm.MITgcm_path — Constant

MITgcm_path

Path to a MITgcm folder. MITgcm_path[1] should generally be used. MITgcm_path[2] is mostly meant to facilitate comparisons between e.g. MITgcm releases when needed.

MITgcm.testreport — Function

testreport(config::MITgcm_config,ext="")

Run the testreport script for one model config, with additional options (optional) speficied in ext

using MITgcm
testreport(MITgcm_config(configuration="front_relax"),"-norun")
#testreport(MITgcm_config(configuration="all"),"-norun")