Model Configurations

MITgcm_config

The data structure that enables MITgcm.jl is called MITgcm_config.

MITgcm.MITgcm_config — Type

MITgcm_config()

Concrete type of AbstractModelConfig for MITgcm (as part of the ClimateModels.jl interface for MITgcm) which contains

    model :: String = "MITgcm"
    configuration :: String = ""
    inputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    outputs :: OrderedDict{Any,Any} = OrderedDict{Any,Any}()
    channel :: Channel{Any} = Channel{Any}(10) 
    folder :: String = tempdir()
    ID :: UUID = UUIDs.uuid4()

and can be constructed using keywords as follows

unknown_config=MITgcm_config(configuration="unknown")

ECCO solutions

Global ocean model configurations used in NASA's ECCO ocean state estimation program.

ECCO4 : Forget et al., 2015 (doi:10.5194/gmd-8-3071-2015)

OCCA2 : Forget, 2024 (under review).

MITgcm.setup_ECCO4! — Function

setup_ECCO4!(config::MITgcm_config)

Setup method for ECCO4 and OCCA2 solutions.

fil=joinpath("examples","configurations","OCCA2.toml")
MC=MITgcm_config(inputs=read_toml(fil))
setup(MC)
build(MC)
MITgcm_launch(MC)

MITgcm/Verification

The MITgcm/verification folder is found in the MITgcm source code. It provides a suite of basic configurations used by model developers.

To use them, you want to download the MITgcm source code. For example :

using MITgcm
MITgcm_download()
MITgcm_path[1]

"/home/runner/.julia/scratchspaces/dce5fa8e-68ce-4431-a242-9469c69627a0/src/MITgcm_test"

To get the list of standard model configurations in MITgcm/verification/ :

verification_experiments()

8-element Vector{MITgcm_config}:
   ID            = 8ac88d83-1aa8-49d0-862b-c2b337c82b9f
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/8ac88d83-1aa8-49d0-862b-c2b337c82b9f
  log subfolder = /tmp/8ac88d83-1aa8-49d0-862b-c2b337c82b9f/log

   ID            = ee85c2d7-5413-4b75-8659-d4273418ede7
  model         = MITgcm
  configuration = advect_cs
  run folder    = /tmp/ee85c2d7-5413-4b75-8659-d4273418ede7
  log subfolder = /tmp/ee85c2d7-5413-4b75-8659-d4273418ede7/log

   ID            = 6d2765b0-5821-489e-8a05-92e6ef80e372
  model         = MITgcm
  configuration = advect_xy
  run folder    = /tmp/6d2765b0-5821-489e-8a05-92e6ef80e372
  log subfolder = /tmp/6d2765b0-5821-489e-8a05-92e6ef80e372/log

   ID            = 1d72fc0e-01ab-4707-908a-3633adebfe47
  model         = MITgcm
  configuration = aim.5l_cs
  run folder    = /tmp/1d72fc0e-01ab-4707-908a-3633adebfe47
  log subfolder = /tmp/1d72fc0e-01ab-4707-908a-3633adebfe47/log

   ID            = 3fa91676-590e-46a2-81a0-76e0fbeff6c2
  model         = MITgcm
  configuration = flt_example
  run folder    = /tmp/3fa91676-590e-46a2-81a0-76e0fbeff6c2
  log subfolder = /tmp/3fa91676-590e-46a2-81a0-76e0fbeff6c2/log

   ID            = b5f93aad-a279-4bd4-af51-1491302b86bb
  model         = MITgcm
  configuration = hs94.cs-32x32x5
  run folder    = /tmp/b5f93aad-a279-4bd4-af51-1491302b86bb
  log subfolder = /tmp/b5f93aad-a279-4bd4-af51-1491302b86bb/log

   ID            = 2ae0a3e4-7bce-41eb-8eb6-ea9f9b0b79ac
  model         = MITgcm
  configuration = tutorial_global_oce_biogeo
  run folder    = /tmp/2ae0a3e4-7bce-41eb-8eb6-ea9f9b0b79ac
  log subfolder = /tmp/2ae0a3e4-7bce-41eb-8eb6-ea9f9b0b79ac/log

   ID            = 06412220-c553-4899-96c7-9ddfdab3fd06
  model         = MITgcm
  configuration = tutorial_held_suarez_cs
  run folder    = /tmp/06412220-c553-4899-96c7-9ddfdab3fd06
  log subfolder = /tmp/06412220-c553-4899-96c7-9ddfdab3fd06/log

To choose one :

MITgcm_config(configuration="MLAdjust")

  ID            = 88668a76-86fd-438b-9938-2e4c29738342
  model         = MITgcm
  configuration = MLAdjust
  run folder    = /tmp/88668a76-86fd-438b-9938-2e4c29738342
  log subfolder = /tmp/88668a76-86fd-438b-9938-2e4c29738342/log

Functionalities

MITgcm.setup_verification! — Function

setup_verification!(config::MITgcm_config)

Setup method for verification experiments.

MITgcm.verification_experiments — Function

verification_experiments()

Get list of all most-standard configurations of MITgcm and return as an Array of MITgcm_config

exps=verification_experiments()

verification_experiments(nam::String)

Get one configurations of MITgcm and return as a MITgcm_config

advect_xy=verification_experiments("advect_xy")

MITgcm.MITgcm_download — Function

MITgcm_download()

Download default, compact version of MITgcm from zenodo.

MITgcm.MITgcm_path — Constant

MITgcm_path

Path to a MITgcm folder. MITgcm_path[1] should generally be used. MITgcm_path[2] is mostly meant to facilitate comparisons between e.g. MITgcm releases when needed.